ALS-11850: Refactoring
