--- pic-sure-resources/pic-sure-ga4gh-dos/src/main/resources/webapp/META-INF/context.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE xml>
-<Context/>
