+++ pic-sure-resources/pic-sure-resource-api/src/main/java/edu/harvard/dbmi/avillach/domain/QueryStatus.java
@@ -3,126 +3,124 @@ package edu.harvard.dbmi.avillach.domain;
 import java.util.Map;
 import java.util.UUID;
 
-import edu.harvard.dbmi.avillach.util.PicSureStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "A query status object")
 public class QueryStatus {
 
-	@Schema(description = "The status of the query", allowableValues = "PENDING, RUNNING, COMPLETED, FAILED, CANCELED")
-	private PicSureStatus status;
+    @Schema(description = "The status of the query", allowableValues = "PENDING, RUNNING, COMPLETED, FAILED, CANCELED")
+    private PicSureStatus status;
 
-	/**
-	 * a uuid associated to a Resource in the database
-	 */
-	@Schema(description = "A UUID associated to a Resource")
-	private UUID resourceID;
+    /**
+     * a uuid associated to a Resource in the database
+     */
+    @Schema(description = "A UUID associated to a Resource")
+    private UUID resourceID;
 
-	@Schema(description = "A status string returned by the resource")
-	private String resourceStatus;
+    @Schema(description = "A status string returned by the resource")
+    private String resourceStatus;
 
-	@Schema(description = "When a user makes a query, a corresponding Result UUID is generated")
-	private UUID picsureResultId;
+    @Schema(description = "When a user makes a query, a corresponding Result UUID is generated")
+    private UUID picsureResultId;
 
-	/**
-	 * when a resource might generate its own resultId and return it,
-	 * we can keep it here
-	 */
-	@Schema(description = "If a resource generates its own resultId, it is stored here")
-	private String resourceResultId;
+    /**
+     * when a resource might generate its own resultId and return it, we can keep it here
+     */
+    @Schema(description = "If a resource generates its own resultId, it is stored here")
+    private String resourceResultId;
 
-	@Schema(description = "The result metadata")
-	private Map<String, Object> resultMetadata;
+    @Schema(description = "The result metadata")
+    private Map<String, Object> resultMetadata;
 
-	@Schema(description = "The size of the result in bytes")
-	private long sizeInBytes;
+    @Schema(description = "The size of the result in bytes")
+    private long sizeInBytes;
 
-	@Schema(description = "The start time of the query")
-	private long startTime;
+    @Schema(description = "The start time of the query")
+    private long startTime;
 
-	@Schema(description = "The duration of the query")
-	private long duration;
+    @Schema(description = "The duration of the query")
+    private long duration;
 
-	@Schema(description = "The expiration time of the query")
-	private long expiration;
+    @Schema(description = "The expiration time of the query")
+    private long expiration;
 
-	public PicSureStatus getStatus() {
-		return status;
-	}
+    public PicSureStatus getStatus() {
+        return status;
+    }
 
-	public void setStatus(PicSureStatus status) {
-		this.status = status;
-	}
+    public void setStatus(PicSureStatus status) {
+        this.status = status;
+    }
 
-	public UUID getResourceID() {
-		return resourceID;
-	}
+    public UUID getResourceID() {
+        return resourceID;
+    }
 
-	public void setResourceID(UUID resourceID) {
-		this.resourceID = resourceID;
-	}
+    public void setResourceID(UUID resourceID) {
+        this.resourceID = resourceID;
+    }
 
-	public String getResourceStatus() {
-		return resourceStatus;
-	}
+    public String getResourceStatus() {
+        return resourceStatus;
+    }
 
-	public void setResourceStatus(String resourceStatus) {
-		this.resourceStatus = resourceStatus;
-	}
+    public void setResourceStatus(String resourceStatus) {
+        this.resourceStatus = resourceStatus;
+    }
 
-	public long getSizeInBytes() {
-		return sizeInBytes;
-	}
+    public long getSizeInBytes() {
+        return sizeInBytes;
+    }
 
-	public void setSizeInBytes(long sizeInBytes) {
-		this.sizeInBytes = sizeInBytes;
-	}
+    public void setSizeInBytes(long sizeInBytes) {
+        this.sizeInBytes = sizeInBytes;
+    }
 
-	public long getStartTime() {
-		return startTime;
-	}
+    public long getStartTime() {
+        return startTime;
+    }
 
-	public void setStartTime(long startTime) {
-		this.startTime = startTime;
-	}
+    public void setStartTime(long startTime) {
+        this.startTime = startTime;
+    }
 
-	public long getDuration() {
-		return duration;
-	}
+    public long getDuration() {
+        return duration;
+    }
 
-	public void setDuration(long duration) {
-		this.duration = duration;
-	}
+    public void setDuration(long duration) {
+        this.duration = duration;
+    }
 
-	public long getExpiration() {
-		return expiration;
-	}
+    public long getExpiration() {
+        return expiration;
+    }
 
-	public void setExpiration(long expiration) {
-		this.expiration = expiration;
-	}
+    public void setExpiration(long expiration) {
+        this.expiration = expiration;
+    }
 
-	public UUID getPicsureResultId() {
-		return picsureResultId;
-	}
+    public UUID getPicsureResultId() {
+        return picsureResultId;
+    }
 
-	public void setPicsureResultId(UUID picsureResultId) {
-		this.picsureResultId = picsureResultId;
-	}
+    public void setPicsureResultId(UUID picsureResultId) {
+        this.picsureResultId = picsureResultId;
+    }
 
-	public String getResourceResultId() {
-		return resourceResultId;
-	}
+    public String getResourceResultId() {
+        return resourceResultId;
+    }
 
-	public void setResourceResultId(String resourceResultId) {
-		this.resourceResultId = resourceResultId;
-	}
+    public void setResourceResultId(String resourceResultId) {
+        this.resourceResultId = resourceResultId;
+    }
 
-	public Map<String, Object> getResultMetadata() {
-		return resultMetadata;
-	}
+    public Map<String, Object> getResultMetadata() {
+        return resultMetadata;
+    }
 
-	public void setResultMetadata(Map<String, Object> resultMetadata) {
-		this.resultMetadata = resultMetadata;
-	}
+    public void setResultMetadata(Map<String, Object> resultMetadata) {
+        this.resultMetadata = resultMetadata;
+    }
 }
