--- pic-sure-util/src/main/java/edu/harvard/dbmi/avillach/util/PicSureStatus.java
@@ -1,8 +1,0 @@
-package edu.harvard.dbmi.avillach.util;
-
-public enum PicSureStatus {
-	QUEUED,
-	PENDING,
-	ERROR,
-	AVAILABLE
-}
